# feat: implement async telemetry with event bus integration