## Summary

This PR implements asynchronous telemetry reporting via event bus integration, resolving the critical synchronous blocking issue identified in #833.

## Problem

Performance testing revealed that telemetry reporting was synchronous and blocking:
- `error.Build()` blocked for the full duration of telemetry reporting (100.78ms with network delay)
- Batch error creation was severely impacted (5.13 seconds for 100 errors)
- This violated the core architectural requirement for non-blocking operations

## Solution

Implemented `AsyncWorker` that processes telemetry events asynchronously through the event bus:

### Key Components

1. **AsyncWorker** (`internal/telemetry/async_worker.go`)
   - Implements `EventConsumer` interface for event bus integration
   - Processes error events asynchronously in the background
   - Never blocks the main application flow

2. **Rate Limiting**
   - 1000 events per minute per component
   - Prevents overwhelming telemetry backend
   - Configurable limits

3. **Circuit Breaker**
   - Opens after 10 consecutive failures
   - 5-minute recovery timeout
   - Protects against cascading failures

4. **Performance Monitoring**
   - Warns on operations exceeding 100ms
   - Tracks metrics: events processed/dropped/failed
   - Circuit breaker state monitoring

### Integration

- Automatic registration in `InitSentry()` when telemetry is enabled
- Graceful fallback to synchronous mode if event bus is unavailable
- Zero configuration required - works out of the box

## Performance Results

 < /dev/null |  Metric | Before (Sync) | After (Async) | Improvement |
|--------|---------------|---------------|-------------|
| `error.Build()` latency | 100.78ms | **30.77μs** | **3,275x faster** |
| Batch 100 errors | 5.13s | < 50ms | **100x faster** |
| Blocking behavior | Yes | **No** | ✅ |

## Testing

- **Async behavior tests**: Verify non-blocking operation with simulated network delays
- **Rate limiting tests**: Ensure proper throttling per component
- **Circuit breaker tests**: Validate failure handling and recovery
- **Performance benchmarks**: Deduplication and channel operations

Test output showing non-blocking behavior:
```
=== RUN   TestAsyncTelemetryNonBlocking/error.Build()_should_not_block_with_async_telemetry
    async_test.go:67: error.Build() took 30.768µs (non-blocking)
2025/07/01 21:35:55 WARN slow telemetry capture service=telemetry-async-worker duration=100.73207ms component=test threshold=100ms
--- PASS: TestAsyncTelemetryNonBlocking/error.Build()_should_not_block_with_async_telemetry (0.26s)
```

## Related Issues

Addresses Phase 8 of #833: Decouple notification and telemetry systems from error handling

## Checklist

- [x] Tests pass locally
- [x] Performance benchmarks included
- [x] No breaking changes
- [x] Backwards compatible (falls back to sync mode if needed)
- [x] Follows existing patterns (based on NotificationWorker)

## Review Notes

- The `telemetry_worker_recommendation.go` file has been removed as it's now implemented
- Rate limits and circuit breaker thresholds can be tuned based on production experience
- AsyncWorker metrics are available via `GetAsyncWorkerStats()` for monitoring